pyproject.toml
[project]
name = "dct-vision"
version = "0.1.0"
description = "Frequency-domain native image processing — skip the pixel decode."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "scipy>=1.15.3",
]

[project.scripts]
dv = "dct_vision.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dct_vision"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]
